Removed coverage setup from Playwright tests and CI
