+++ .github/workflows/unit_tests.yaml
@@ -26,13 +26,4 @@ jobs:
         run: npm install
 
       - name: Run vitest tests
-        run: npx vitest run --coverage.enabled --coverage.provider=istanbul --coverage.include=src --coverage.reporter=json --coverage.all
-
-      - uses: actions/upload-artifact@v3
-        with:
-          name: unit-coverage
-          path: coverage-unit/
-          retention-days: 30
-
-    env:
-      ENABLE_COVERAGE: "true"
+        run: npx vitest run
