--- .nycrc
@@ -1,3 +1,0 @@
-{
-  "all": true
-}
