+++ docs/development/tests.md
@@ -18,24 +18,15 @@ To execute the tests seeing the browser add the `--headed` flag or the `--debug`
 
 ## Coverage
 
+> Warning: code coverage results are not reliable at the moment
+
 Coverage for the unit tests:
 
 ```
-ENABLE_COVERAGE=true npx vitest run --coverage.enabled --coverage.provider=istanbul --coverage.include=src --coverage.reporter=json --coverage.all
+npx vitest --coverage
 ```
 
-Warning: coverage takes some time using the option `--coverage.all`.
-
-Coverage for the playwright tests is automatically collected when running `ENABLE_COVERAGE=true npx playwright test`.
-
-To generate a merged html report:
-
-```
-mkdir coverage-all
-cp coverage-playwright/* coverage-all
-cp coverage-unit/* coverage-all
-npx nyc report --report-dir ./coverage-html --temp-dir ./coverage-all --reporter=html --exclude-after-remap false
-```
+Warning: coverage takes some time, since we are using the option `{ all: true }`.
 
 ### Local `fractal-server` instance
