+++ tests/collect_core_tasks.setup.js
@@ -1,4 +1,5 @@
-import { expect, test, waitPageLoading } from './base_test.js';
+import { expect, test } from '@playwright/test';
+import { waitPageLoading } from './utils.js';
 
 test('Collect core tasks', async ({ page }) => {
 	await page.goto('/tasks');
