+++ tests/profile.spec.js
@@ -1,4 +1,5 @@
-import { expect, test, waitPageLoading } from './base_test.js';
+import { expect, test } from '@playwright/test';
+import { waitPageLoading } from './utils.js';
 
 test('User profile', async ({ page }) => {
 	await test.step('Open user profile page', async () => {
